server/requirements.txt
pyngrok
runway-python
scikit-image
torch
torchvision